changing some things back to size_t
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5498 d7cf8633-e32d-0410-b094-e92efae38249

+++ snd_main.h
@@ -26,8 +26,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 typedef struct
 {
-	int		length;
-	int		offset;
+	size_t	length;
+	size_t	offset;
 	qbyte	data[4];	// variable sized
 } sfxbuffer_t;
 
@@ -59,7 +59,7 @@ struct sfx_s
 	unsigned int		flags;			// cf SFXFLAG_* defines
 	snd_format_t		format;
 	int					loopstart;
-	int					total_length;
+	size_t				total_length;
 	const snd_fetcher_t	*fetcher;
 	void				*fetcher_data;	// Per-sfx data for the sound fetching functions
 };
